Refactor circuit cutting into gate and wire cutting
